LayoutTests/fast/forms/select/select-disabled-expected.txt

Tests that user cannot change disabled select by keyboard

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select1.value is "a"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select2.value is "a"
PASS select2.value is "b"
PASS select2.value is "b"
PASS select2.value is "b"
PASS successfullyParsed is true

TEST COMPLETE
